feat(x/staking): `A-B-C-D-E, F-U`; And your mom and your sister and your job; And your craigslist couch and the way your voice sounds (#588)
* chore(repo): remove deadcode

* validator type

* godoc

* lint

* push

* bet

* runs somehow

* runs

* bet

* bet

* what staking

* chain run

* bet

* reintro randao properly

* make even smaller

* houdini

* fix make-devnet

* more

* forgot

* bet

* linter

* hook init

* bet

* should pass tests now

* bada boom bada bing

* skip for now

* reduce replaces

* more replaces reduction

* bet

* we ride

* bet

* cleanup

* bet

* bet

* fix bug

* bet

* feat(types): Test DepositCredentials

* lint

* bet

* bet

* bump solidity

* fix

* e2e

* should pass ci

* remove consensus address usage

* remove consensus address usage

* organize module

* lint

* tidy

* linter

* typo

* cleanup imports

* move folders

* clean

* add validators to state

* gen state

* fix genesis

* bet

* linter

* ON GOD FR FR

* bet

* remove unused

* remove deadcode

* address ocnc comments

* readd-nether

---------

+++ e2e/suite/constants.go
@@ -35,8 +35,8 @@ import (
 // transactions.
 const (
 	Ether   = params.Ether
-	OneGwei = params.GWei  // 1 Gwei = 1e9 wei
-	TenGwei = 10 * OneGwei // 10 Gwei = 1e10 wei
+	OneGwei = uint64(params.GWei) // 1 Gwei = 1e9 wei
+	TenGwei = 10 * OneGwei        // 10 Gwei = 1e10 wei
 )
 
 // EtherTransferGasLimit specifies the gas limit for a standard Ethereum
